Fix py_test
